Revert "[Tests] small test code refactor (#121)"
This reverts commit 37d87e6e18af488dd87563c192f4d518a8611935.
