+++ chain-impl-mockchain/src/leadership/genesis/mod.rs
@@ -551,10 +551,8 @@ mod tests {
             .is_err());
     }
 
-    use crate::block::{Block, Header};
     use crate::fragment::Contents;
     use crate::header::{BlockVersion, HeaderBuilderNew};
-    use crate::testing::data::StakePool;
     use chain_core::property::ChainLength;
 
     #[test]
