+++ chain-impl-mockchain/src/ledger/ledger.rs
@@ -958,10 +958,9 @@ mod tests {
         setting::Settings,
         testing::{
             address::ArbitraryAddressDataValueVec,
-            builders::{create_initial_transaction, witness_builder, TransactionBuilder},
+            builders::{witness_builder, TransactionBuilder},
             data::{AddressData, AddressDataValue},
             ledger::{self as ledger_mock, ConfigBuilder},
-            verifiers::LedgerStateVerifier,
             TestGen,
         },
         transaction::Witness,
@@ -1221,8 +1220,10 @@ mod tests {
         witness: Witness,
     ) -> TestResult {
         let faucet = AddressData::utxo(Discrimination::Test);
-        let message =
-            create_initial_transaction(Output::from_address(faucet.address.clone(), Value(100)));
+        let message = ledger_mock::create_initial_transaction(Output::from_address(
+            faucet.address.clone(),
+            Value(100),
+        ));
         let (_, ledger) =
             ledger_mock::create_initial_fake_ledger(&[message], ConfigBuilder::new().build())
                 .unwrap();
@@ -1243,8 +1244,10 @@ mod tests {
     #[test]
     pub fn test_input_utxo_verify_correct_utxo() {
         let faucet = AddressData::utxo(Discrimination::Test);
-        let message =
-            create_initial_transaction(Output::from_address(faucet.address.clone(), Value(100)));
+        let message = ledger_mock::create_initial_transaction(Output::from_address(
+            faucet.address.clone(),
+            Value(100),
+        ));
         let (block0_hash, ledger) =
             ledger_mock::create_initial_fake_ledger(&[message], ConfigBuilder::new().build())
                 .unwrap();
@@ -1267,8 +1270,10 @@ mod tests {
     #[test]
     pub fn test_input_utxo_verify_incorrect_value() {
         let faucet = AddressData::utxo(Discrimination::Test);
-        let message =
-            create_initial_transaction(Output::from_address(faucet.address.clone(), Value(100)));
+        let message = ledger_mock::create_initial_transaction(Output::from_address(
+            faucet.address.clone(),
+            Value(100),
+        ));
         let (block0_hash, ledger) =
             ledger_mock::create_initial_fake_ledger(&[message], ConfigBuilder::new().build())
                 .unwrap();
@@ -1420,6 +1425,121 @@ mod tests {
                 output: output,
             };
             utxo
+        }
+    }
+
+    pub struct LedgerStateVerifier {
+        multisig_ledger: multisig::Ledger,
+        utxos: utxo::Ledger<Address>,
+        accounts: account::Ledger,
+    }
+
+    impl LedgerStateVerifier {
+        pub fn new(ledger: Ledger) -> Self {
+            LedgerStateVerifier {
+                multisig_ledger: ledger.multisig,
+                utxos: ledger.utxos,
+                accounts: ledger.accounts,
+            }
+        }
+
+        pub fn utxo_contains(&self, entry: &utxo::Entry<Address>) -> &Self {
+            assert_eq!(
+                self.utxos.iter().find(|x| *x == entry.clone()),
+                Some(entry.clone())
+            );
+            self
+        }
+
+        pub fn and(&self) -> &Self {
+            self
+        }
+
+        pub fn accounts_contains(
+            &self,
+            id: Identifier,
+            expected_account_state: AccountState<()>,
+        ) -> &Self {
+            let account_state = self.accounts.get_state(&id).unwrap();
+            assert_eq!(account_state.clone(), expected_account_state);
+            self
+        }
+
+        pub fn utxos_count_is(&self, count: usize) -> &Self {
+            assert_eq!(
+                self.utxos.iter().count(),
+                count,
+                "Utxo count should be equal to {:?}",
+                count
+            );
+            self
+        }
+
+        pub fn accounts_count_is(&self, count: usize) -> &Self {
+            assert_eq!(
+                self.accounts.iter().count(),
+                count,
+                "Utxo count should be equal to {:?}",
+                count
+            );
+            self
+        }
+
+        pub fn multisigs_count_is_zero(&self) -> &Self {
+            assert_eq!(self.multisig_ledger.iter_accounts().count(), 0);
+            assert_eq!(self.multisig_ledger.iter_declarations().count(), 0);
+            self
+        }
+
+        pub fn total_value_is(&self, value: Value) -> &Self {
+            let account_total = self.accounts.get_total_value().unwrap();
+            let multisig_total = self.multisig_ledger.get_total_value().unwrap();
+            let utxos_total =
+                Value::sum(self.utxos.iter().map(|entry| entry.output.value)).unwrap();
+            let totals = vec![account_total, multisig_total, utxos_total];
+            let actual_value =
+                Value::sum(totals.iter().cloned()).expect("cannot sum up ledger total value");
+            assert_eq!(
+                value, actual_value,
+                "Expected value {:?} vs {:?} of actual",
+                value, actual_value
+            );
+            self
+        }
+
+        // Does not cover situation in which we have two identical utxos
+        pub fn address_has_expected_balance(&self, address: AddressData, value: Value) -> &Self {
+            match self.accounts.exists(&address.to_id()) {
+                true => self.account_has_expected_balance(address, value),
+                false => self.utxo_has_expected_balance(address, value),
+            }
+        }
+
+        pub fn account_has_expected_balance(&self, address: AddressData, value: Value) -> &Self {
+            let account_state = self
+                .accounts
+                .get_state(&address.to_id())
+                .expect("account does not exists while it should");
+            assert_eq!(account_state.value(), value);
+            self
+        }
+
+        pub fn utxo_has_expected_balance(&self, address_data: AddressData, value: Value) -> &Self {
+            let utxo = self
+                .utxos
+                .iter()
+                .find(|x| *x.output == address_data.make_output(value));
+            match value == Value::zero() {
+                true => {
+                    assert!(utxo.is_none());
+                    return self;
+                }
+                false => {
+                    let utxo = utxo.unwrap();
+                    assert_eq!(utxo.output.value, value);
+                    return self;
+                }
+            }
         }
     }
 
@@ -1726,7 +1846,7 @@ mod tests {
             fees: LinearFee::new(100, 0, 0),
             reward_params: Some(RewardParams::Linear(0, 0, 0)),
         };
-        let mut faucet = AddressDataValue::account(Discrimination::Test, Value(101));
+        let faucet = AddressDataValue::account(Discrimination::Test, Value(101));
         let reciever = AddressDataValue::account(Discrimination::Test, Value(1));
         let (block0_hash, ledger) = ledger_mock::create_fake_ledger_with_faucet(
             &[faucet.clone()],
