+++ chain-impl-mockchain/src/ledger/tests/ledger_tests.rs
@@ -15,7 +15,7 @@ use crate::{
         },
         data::AddressData,
         ledger::{self, ConfigBuilder},
-        builders,
+        requests,
         tx_builder::TransactionBuilder,
         TestGen,
     },
@@ -54,7 +54,7 @@ pub fn ledger_accepts_correct_transaction(
     receiver: AddressData,
     value: NonZeroValue,
 ) -> TestResult {
-    let message = builders::create_initial_transaction(Output::from_address(
+    let message = requests::create_initial_transaction(Output::from_address(
         faucet.address.clone(),
         value.into(),
     ));
@@ -100,7 +100,7 @@ pub fn total_funds_are_const_in_ledger(
     mut transaction_data: ArbitraryValidTransactionData,
 ) -> TestResult {
     let message =
-        builders::create_initial_transactions(&transaction_data.make_outputs_from_all_addresses());
+        requests::create_initial_transactions(&transaction_data.make_outputs_from_all_addresses());
     let (block0_hash, ledger) = ledger::create_initial_fake_ledger(
         &[message],
         ConfigBuilder::new()
@@ -169,7 +169,7 @@ pub fn utxo_no_enough_signatures() {
     let faucet = AddressData::utxo(Discrimination::Test);
     let receiver = AddressData::utxo(Discrimination::Test);
 
-    let message = builders::create_initial_transaction(Output::from_address(
+    let message = requests::create_initial_transaction(Output::from_address(
         faucet.address.clone(),
         Value(42000),
     ));
@@ -204,7 +204,7 @@ pub fn transaction_with_more_than_253_outputs() {
         outputs.push(Output::from_address(receiver.address.clone(), Value(1)));
     }
 
-    let message = builders::create_initial_transaction(Output::from_address(
+    let message = requests::create_initial_transaction(Output::from_address(
         faucet.address.clone(),
         Value(256),
     ));
@@ -236,7 +236,7 @@ pub fn transaction_with_more_than_253_outputs() {
 pub fn iterate() {
     let faucet = AddressData::utxo(Discrimination::Test);
 
-    let message = builders::create_initial_transaction(Output::from_address(
+    let message = requests::create_initial_transaction(Output::from_address(
         faucet.address.clone(),
         Value(42000),
     ));
