+++ chain-impl-mockchain/src/testing/builders/mod.rs
@@ -7,11 +7,9 @@ pub mod tx_builder;
 pub mod tx_cert_builder;
 pub mod witness_builder;
 pub mod update_builder;
-mod initial_builder;
 
 pub use block_builder::*;
 pub use cert_builder::*;
-pub use initial_builder::*;
 pub use stake_pool_builder::*;
 pub use old_address_builder::*;
 pub use proposal_builder::*;pub use tx_builder::*;
