+++ chain-impl-mockchain/src/testing/builders/tx_cert_builder.rs
@@ -87,13 +87,6 @@ impl TransactionCertAuthenticator {
         self
     }
 
-    pub fn with_witnesses(&mut self, block0: &HeaderId, addresses_data: &Vec<AddressData>) -> &mut Self {
-        for address_data in addresses_data {
-            self.with_witness(block0,address_data);
-        }
-        self
-    }
-
     pub fn hash(&self) -> TransactionSignDataHash {
         txbuilder::TransactionFinalizer::new(self.transaction.clone()).get_tx_sign_data_hash()
     }
