+++ chain-impl-mockchain/src/testing/data/address.rs
@@ -217,53 +217,44 @@ impl AddressDataValue {
     pub fn new(address_data: AddressData, value: Value) -> Self {
         AddressDataValue {
             address_data: address_data,
-            value: value
+            value: value,
         }
     }
 
     pub fn utxo(discrimination: Discrimination, value: Value) -> Self {
-        AddressDataValue::new(AddressData::utxo(discrimination),value)
+        AddressDataValue {
+            address_data: AddressData::utxo(discrimination),
+            value: value,
+        }
     }
 
     pub fn account(discrimination: Discrimination, value: Value) -> Self {
-        AddressDataValue::new(AddressData::account(discrimination),value)
+        AddressDataValue {
+            address_data: AddressData::account(discrimination),
+            value: value,
+        }
     }
 
     pub fn delegation(discrimination: Discrimination, value: Value) -> Self {
-        AddressDataValue::new(AddressData::delegation(discrimination),value)
+        AddressDataValue {
+            address_data: AddressData::delegation(discrimination),
+            value: value,
+        }
     }
 
     pub fn to_id(&self) -> Identifier {
         self.address_data.to_id()
     }
 
-    pub fn public_key(&self) -> PublicKey<Ed25519> {
-        self.address_data.public_key()
-    }
-
     pub fn private_key(&self) -> EitherEd25519SecretKey {
         self.address_data.private_key.clone()
     }
-
     pub fn make_input(&self, utxo: Option<Entry<Address>>) -> Input {
-        self.make_input_with_value(utxo,self.value)
-    }
-
-    pub fn make_input_with_value(&self, utxo: Option<Entry<Address>>, value: Value) -> Input {
-        self.address_data.make_input(value, utxo)
+        self.address_data.make_input(self.value, utxo)
     }
 
     pub fn make_output(&self) -> Output<Address> {
-        self.make_output_with_value(self.value)
-    }
-
-    pub fn make_output_with_value(&self, value: Value) -> Output<Address> {
-        self.address_data.make_output(value)
-    }
-
-    pub fn increment_spending_counter(&mut self) {
-        let counter: u32 = self.address_data.spending_counter.unwrap().into();
-        self.address_data.spending_counter = Some((counter + 1u32).into());
+        self.address_data.make_output(self.value)
     }
 }
