+++ chain-impl-mockchain/src/testing/data/mod.rs
@@ -1,9 +1,7 @@
-mod address;
-mod leader;
-mod stake_pool;
-mod wallet;
+pub mod address;
+pub mod leader;
+pub mod stake_pool;
 
 pub use address::*;
 pub use leader::*;
 pub use stake_pool::*;
-pub use wallet::*;
