--- chain-impl-mockchain/src/testing/verifiers/mod.rs
@@ -1,3 +1,0 @@
-mod ledger_verifier;
-
-pub use ledger_verifier::LedgerStateVerifier;
